[FEAT] : get book list
