+++ src/router/index.ts
@@ -1,8 +1,7 @@
-import { Router } from "express";
-import userRouter from "./userRouter";
-
+import { Router } from 'express';
+import bookRouter from './bookRouter';
 const router: Router = Router();
 
-router.use("/user", userRouter);
+router.use('/book', bookRouter);
 
 export default router;
